requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Environment
python-dotenv==1.0.0

# File Upload
python-multipart==0.0.6

# Utilities
pydantic==2.5.0